part3 debug

+++ sieve3.c
@@ -98,8 +98,8 @@ int main (int argc, char *argv[])
    for (i = 0; i < size; i++) marked[i] = 0;
 
    int block_size = 8;
-   unsigned long int block_low_value = low_value;
-   unsigned long int block_high_value = block_low_value + 2 * (block_size - 1);
+   unsigned long long int block_low_value = low_value;
+   unsigned long long int block_high_value = block_low_value + 2 * (block_size - 1);
    do {
       index = 0;
       prime = 3;
